Performed Swift migration to Swift 3.
Basically just ran the migrator and fixed the few errors that resulted.
Ran the test suite after migration, still passing.

There was one place where I intervened. The migrator found a value that
it knew needed to conform to NSCopying and NSObjectProtocol...
`context.setObject(self, forKeyedSubscript: "$console" as (NSCopying &
NSObjectProtocol)!)` I changed this to be `as NSString`